refactor: provider path name update
