+++ app/(tabs)/index.tsx
@@ -1,7 +1,7 @@
 import UserHome from "@/views/UserHome";
 import LoginPage from "@/views/LoginPage";
 import PageWrapper from "@/views/components/PageWrapper";
-import { AuthContext } from '@/context/AuthProvider';
+import { AuthContext } from '@/providers/AuthProvider';
 import {ReactNode, useContext} from "react";
 
 export default function Index(): ReactNode {
